+++ scripts/inventory_scripts/remove_inventory_item.php
@@ -1,19 +1,22 @@
 <?php
-function removeInventoryItem( $inventoryID) {
     // Include the config file
-    require_once '../data/config.php';
+    require '../data/config.php';
 
-    // remove the inventory item data from the database
-    $query = "DELETE FROM inventory WHERE inventoryID = $inventoryID";
-    $result = $conn->query($query);
+    if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+        if (isset($_POST['delete-product-id'])) {
+            $inventoryID = $_POST['delete-product-id'];
 
-    // Check if the data was removed successfully
-    if ($result) {
-        echo "Inventory item removed successfully.";
-    } else {
-        echo "Error removing inventory item: " . $conn->error;
+            // remove the inventory item data from the database
+            $query = "DELETE FROM inventory WHERE inventoryID = $inventoryID";
+            $result = $conn->query($query);
+
+            // Check if the data was removed successfully
+            if ($result) {
+                echo "<script> console.log('Inventory item removed successfully.')</script>";
+            } else {
+                echo "<script> console.log('Error removing inventory item: " . $conn->error. "')</script>"; 
+            }
+        }
     }
-
-    // Close the database connection
-    $conn->close();
-}
+        // Close the database connection
+        $conn->close();
